board name
보드네임을 추가하였습니다.

+++ GSupport/src/main/java/com/team_c/dao/LoginDAO.java
@@ -17,7 +17,7 @@ public class LoginDAO extends AbstractDAO{
 	}
 	
 	public int emailCheck(String email) {
-		return selectOne("login.emailCheck", email);
+		return (int) selectOne("login.emailCheck", email);
 	}
 	
 }
